pyproject.toml
[project]
name = "dayana-apto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0.0",
    "dj-database-url>=3.0.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.11.0",
    "psycopg>=3.1.0",
    "djangorestframework-simplejwt>=5.5.1",
    "django-allauth>=65.13.1",
    "dj-rest-auth[with-social]>=7.0.1",
    "django-money>=3.5.4",
    "django-filter>=25.2",
    "factory-boy>=3.3.3",
    "pytest-django>=4.11.1",
]
